Not sure if we can say respectively here
Testing too early certainly can lead to false negatives.

Testing late could also be subject to false negatives. That would
explain the pattern of Positive-Negative-Positive that happens in rare
occasions. But that pattern could equally be explained by the subject
being a true negative who received a false positive afterwards if they
didn't try culturing it to see if it was viable

(I think)