I think this is bullshit
The cytokine storm theory is floated around a lot, but it doesn't
necessarily explain that mortality pattern. Indeed this one is probably better
described the all-cause mortality curve.

Note that pandemic flu is the best flu to look at as far as outbreaks of
novel viruses. Seasonal flu dynamics are different. And the large young
adult death presence seen in pandemic flu itself may be also confounded
by the conditions those same young adults were in at the time.